+++ fundamentals/today-i-learned/src/App.tsx
@@ -1,9 +1,9 @@
 import { Route, Routes } from "react-router-dom";
 import { Layout } from "./components/shared/layout/Layout";
-import { NewHomePage } from "./pages/newHome";
+import { LegacyTimelinePage } from "./pages/legacy-timeline/TimelinePage";
+import { PostDetailPage } from "./pages/postDetail/PostDetailPage";
 import { MyPage } from "./pages/profile/MyPage";
 import { TimelinePage } from "./pages/timeline/TimelinePage";
-import { PostDetailPage } from "./pages/postDetail/PostDetailPage";
 
 function App() {
   return (
@@ -11,7 +11,7 @@ function App() {
       <Routes>
         <Route path="/" element={<TimelinePage />} />
         <Route path="/profile" element={<MyPage />} />
-        <Route path="/new-home" element={<NewHomePage />} />
+        <Route path="/legacy-timeline" element={<LegacyTimelinePage />} />
         <Route path="/post/:id" element={<PostDetailPage />} />
       </Routes>
     </Layout>
