+++ fundamentals/today-i-learned/src/components/shared/ui/Card.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { cva, type VariantProps } from "class-variance-authority";
-import { cn } from "@/lib/utils/cn";
+import { cn } from "@/libs/cn";
 
 const cardVariants = cva("rounded-2xl border bg-white transition-all", {
   variants: {
