+++ fundamentals/today-i-learned/src/pages/legacy-timeline/components/CreatePost.tsx
