+++ fundamentals/today-i-learned/src/pages/postDetail/components/PostContent.tsx
@@ -1,4 +1,4 @@
-import type { Post } from "../../newHome/utils/types";
+import type { Post } from "../../timeline/utils/types";
 
 interface PostContentProps {
   post: Post;
@@ -12,12 +12,10 @@ export function PostContent({ post }: PostContentProps) {
       return "bg-[rgba(237,204,248,0.4)] text-[#DA9BEF]";
     if (tagName.includes("가독성"))
       return "bg-[rgba(255,212,214,0.4)] text-[#FB8890]";
-    if (tagName.includes("결합도"))
-      return "bg-green-100 text-green-700";
+    if (tagName.includes("결합도")) return "bg-green-100 text-green-700";
     if (tagName.includes("기여") || tagName.includes("설계"))
       return "bg-orange-100 text-orange-700";
-    if (tagName.includes("Changes"))
-      return "bg-yellow-100 text-yellow-700";
+    if (tagName.includes("Changes")) return "bg-yellow-100 text-yellow-700";
     if (tagName.includes("조건부") || tagName.includes("렌더링"))
       return "bg-blue-100 text-blue-700";
     return "bg-gray-100 text-gray-600";
