+++ fundamentals/today-i-learned/src/pages/postDetail/hooks/usePostDetail.ts
@@ -1,22 +1,25 @@
 import { useState } from "react";
-import { 
-  useDiscussionDetail, 
+import {
+  useDiscussionDetail,
   useWeeklyTopDiscussions,
   useAddDiscussionComment,
   useToggleDiscussionReaction
 } from "@/api/hooks/useDiscussions";
-import type { 
-  GitHubDiscussionDetail, 
-  GitHubComment 
+import type {
+  GitHubDiscussionDetail,
+  GitHubComment
 } from "@/api/remote/discussions";
-import type { Post, Comment, PopularPost } from "../../newHome/utils/types";
+import type { Post, Comment, PopularPost } from "../../timeline/utils/types";
 
 function truncateText(text: string, maxLength: number): string {
   if (text.length <= maxLength) return text;
   return text.slice(0, maxLength) + "...";
 }
 
-function mapGitHubUserToAuthor(githubUser: { login: string; avatarUrl: string }) {
+function mapGitHubUserToAuthor(githubUser: {
+  login: string;
+  avatarUrl: string;
+}) {
   return {
     id: githubUser.login,
     name: githubUser.login,
@@ -35,17 +38,18 @@ function mapGitHubCommentToComment(comment: GitHubComment): Comment {
       upvotes: comment.reactions.totalCount,
       replies: comment.replies?.totalCount || 0
     },
-    replies: comment.replies?.nodes?.map(reply => ({
-      id: reply.id,
-      content: reply.body,
-      author: mapGitHubUserToAuthor(reply.author),
-      createdAt: reply.createdAt,
-      stats: {
-        upvotes: reply.reactions.totalCount,
-        replies: 0
-      },
-      parentId: comment.id
-    })) || []
+    replies:
+      comment.replies?.nodes?.map((reply) => ({
+        id: reply.id,
+        content: reply.body,
+        author: mapGitHubUserToAuthor(reply.author),
+        createdAt: reply.createdAt,
+        stats: {
+          upvotes: reply.reactions.totalCount,
+          replies: 0
+        },
+        parentId: comment.id
+      })) || []
   };
 }
 
@@ -57,7 +61,7 @@ function mapDiscussionToPost(discussion: GitHubDiscussionDetail): Post {
     author: mapGitHubUserToAuthor(discussion.author),
     createdAt: discussion.createdAt,
     category: discussion.category.name,
-    tags: discussion.labels?.nodes?.map(label => label.name) || [],
+    tags: discussion.labels?.nodes?.map((label) => label.name) || [],
     stats: {
       upvotes: discussion.reactions.totalCount,
       hearts: 0,
@@ -69,7 +73,11 @@ function mapDiscussionToPost(discussion: GitHubDiscussionDetail): Post {
 }
 
 export function usePostDetail(postId: string | undefined) {
-  const { data: discussion, isLoading, error } = useDiscussionDetail(postId || "");
+  const {
+    data: discussion,
+    isLoading,
+    error
+  } = useDiscussionDetail(postId || "");
   const { data: weeklyPosts } = useWeeklyTopDiscussions({ limit: 5 });
   const addComment = useAddDiscussionComment();
   const toggleReaction = useToggleDiscussionReaction();
@@ -81,13 +89,13 @@ export function usePostDetail(postId: string | undefined) {
 
   const handleLike = (postId: string) => {
     const isLiked = interactionState.likes.has(postId);
-    toggleReaction.mutate({ 
-      subjectId: postId, 
+    toggleReaction.mutate({
+      subjectId: postId,
       isReacted: isLiked,
       content: "HEART"
     });
-    
-    setInteractionState(prev => {
+
+    setInteractionState((prev) => {
       const newLikes = new Set(prev.likes);
       if (newLikes.has(postId)) {
         newLikes.delete(postId);
@@ -108,13 +116,13 @@ export function usePostDetail(postId: string | undefined) {
 
   const handleUpvote = (postId: string) => {
     const isUpvoted = interactionState.upvotes.has(postId);
-    toggleReaction.mutate({ 
-      subjectId: postId, 
+    toggleReaction.mutate({
+      subjectId: postId,
       isReacted: isUpvoted,
       content: "THUMBS_UP"
     });
-    
-    setInteractionState(prev => {
+
+    setInteractionState((prev) => {
       const newUpvotes = new Set(prev.upvotes);
       if (newUpvotes.has(postId)) {
         newUpvotes.delete(postId);
@@ -126,8 +134,8 @@ export function usePostDetail(postId: string | undefined) {
   };
 
   const handleCommentUpvote = (commentId: string) => {
-    toggleReaction.mutate({ 
-      subjectId: commentId, 
+    toggleReaction.mutate({
+      subjectId: commentId,
       isReacted: false,
       content: "THUMBS_UP"
     });
@@ -135,12 +143,12 @@ export function usePostDetail(postId: string | undefined) {
 
   const handleReply = (commentId: string, content: string) => {
     if (!discussion?.id || !content.trim()) return;
-    
-    const parentComment = comments.find(comment => comment.id === commentId);
-    const replyPrefix = parentComment 
+
+    const parentComment = comments.find((comment) => comment.id === commentId);
+    const replyPrefix = parentComment
       ? `> @${parentComment.author.name}: ${content}\n\n`
       : `> ${content}\n\n`;
-    
+
     addComment.mutate({
       discussionId: discussion.id,
       body: `${replyPrefix}답글 내용을 입력하세요.`
@@ -149,7 +157,7 @@ export function usePostDetail(postId: string | undefined) {
 
   const handleNewComment = (content: string) => {
     if (!discussion?.id || !content.trim()) return;
-    
+
     addComment.mutate({
       discussionId: discussion.id,
       body: content
@@ -157,17 +165,20 @@ export function usePostDetail(postId: string | undefined) {
   };
 
   const postData = discussion ? mapDiscussionToPost(discussion) : null;
-  
-  const comments = discussion?.comments.nodes.map(mapGitHubCommentToComment) || [];
+
+  const comments =
+    discussion?.comments.nodes.map(mapGitHubCommentToComment) || [];
 
   const weeklyTop5Data = {
-    posts: (weeklyPosts || []).map((post, index): PopularPost => ({
-      id: post.id,
-      title: post.title,
-      author: mapGitHubUserToAuthor(post.author),
-      excerpt: truncateText(post.body.replace(/[#*`\n]/g, " ").trim(), 80),
-      rank: index + 1
-    })),
+    posts: (weeklyPosts || []).map(
+      (post, index): PopularPost => ({
+        id: post.id,
+        title: post.title,
+        author: mapGitHubUserToAuthor(post.author),
+        excerpt: truncateText(post.body.replace(/[#*`\n]/g, " ").trim(), 80),
+        rank: index + 1
+      })
+    ),
     weekInfo: "이번 주 인기글"
   };
 
@@ -175,11 +186,11 @@ export function usePostDetail(postId: string | undefined) {
     postId,
     isLoading,
     error,
-    
+
     postData,
     comments,
     weeklyTop5Data,
-    
+
     handlers: {
       handleLike,
       handleComment,
