+++ fundamentals/today-i-learned/src/pages/timeline/utils/types.ts
